3percent rule / prioritize covered incorrect percentage displayed bug fix.
