+++ iacs_qa/sample_extraction/JRC_v2_notebook/.gitignore
@@ -2,6 +2,9 @@
 *.csv
 input/
 output/
+tests/
 *.pickle
+log.txt
 test_notebook.ipynb
+compare_outputs.py
 anonymizer.py
