+++ iacs_qa/sample_extraction/JRC_v2_notebook/modules/input_verification.py
@@ -6,73 +6,14 @@ class DataFrameValidationError(Exception):
     """Exception raised for errors in the input DataFrame."""
     pass
 
-def verify_parcel_df(parcel_df):
-    """
-    Verifies if all required columns are present in the dataframe, their data types are correct,
-    and there are no empty values.
-
-    Raises:
-        DataFrameValidationError: If any verification check fails.
-    
-    Returns:
-        True if the dataframe is valid.
-    """
-    requirements = {
-        "gsa_par_id": [("string", "string"), ("int64", "integer")],
-        "gsa_hol_id": [("int64", "integer"),],
-        "ua_grp_id": [("string", "string")],
-        "covered": [("int64", "integer")],
-        "ranking": [("int64", "integer")]
-    }
-
-    for column, expected_types in requirements.items():
-        if column not in parcel_df.columns:
-            raise DataFrameValidationError(f"Column '{column}' not found in the dataframe.")
-        if parcel_df[column].isnull().any():
-            raise DataFrameValidationError(f"Column '{column}' contains empty values. First line with empty value: {parcel_df[parcel_df[column].isnull()].index[0] + 2}")
-        
-        actual_type = parcel_df[column].dtype
-
-        # Special check for the 'covered' column to ensure it's binary (0 or 1)
-        if column == "covered":
-            if not pd.api.types.is_integer_dtype(parcel_df[column]):
-                raise DataFrameValidationError(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
-            if not parcel_df[column].isin([0, 1]).all():
-                raise DataFrameValidationError(f"Column '{column}' should only contain 0s or 1s.")
-        else:
-            type_matched = False
-            for expected_type_str, type_name in expected_types:
-                if expected_type_str == "string":
-                    if pd.api.types.is_string_dtype(parcel_df[column]) or actual_type == "object":
-                        type_matched = True
-                        break
-                else:
-                    if pd.api.types.is_dtype_equal(actual_type, expected_type_str):
-                        type_matched = True
-                        break
-            
-            if not type_matched:
-                raise DataFrameValidationError(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
-
-    return True
-
 def verify_and_report_parcel_df(parcel_df):
     """
-    Verifies if all required columns are present in the dataframe, their data types are correct,
-    and there are no empty values.
-
-    Raises:
-        DataFrameValidationError: If any verification check fails. Lists all issues found in the dataframe.
-    
-    Returns:
-        True if the dataframe is valid.
-
-    To be implemented: a text file listing all incorrect rows.
+    Fix this. This should collect all problems and throw them as a single exception.
     """
 
     requirements = {
-        "gsa_par_id": [("string", "string"), ("int64", "integer"), ("float64", "float")],
-        "gsa_hol_id": [("int64", "integer"), ("float64", "float")],
+        "gsa_par_id": [("string", "string"), ("int64", "integer"), ("float64", "float"),],
+        "gsa_hol_id": [("int64", "integer"), ("float64", "float"),("string", "string"),],
         "ua_grp_id": [("string", "string")],
         "covered": [("int64", "integer")],
         "ranking": [("int64", "integer")]
@@ -80,46 +21,57 @@ def verify_and_report_parcel_df(parcel_df):
 
     issues_found = []
 
-    empty_values = False
-    covered_wrong_values = False
-
-
     for column, expected_types in requirements.items():
+        # Check if column exists
         if column not in parcel_df.columns:
             issues_found.append(f"Column '{column}' not found in the dataframe.")
+            continue
 
+        # Check for empty values
         try:
             if parcel_df[column].isnull().any():
                 issues_found.append(f"Column '{column}' contains empty values.")
-        except:
-            issues_found.append(f"Column '{column}' contains empty values.")    
+        except Exception as e:
+            issues_found.append(f"Error checking empty values in column '{column}': {str(e)}")
 
-        actual_type = parcel_df[column].dtype
+        # Get actual type
+        try:
+            actual_type = parcel_df[column].dtype
+        except Exception as e:
+            issues_found.append(f"Error getting data type of column '{column}': {str(e)}")
+            continue
 
-        # Special check for the 'covered' column to ensure it's binary (0 or 1)
+        # Special check for the 'covered' column
         if column == "covered":
-            if not pd.api.types.is_integer_dtype(parcel_df[column]):
-                issues_found.append(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
-            if not parcel_df[column].isin([0, 1]).all():
-                issues_found.append(f"Column '{column}' should only contain 0s or 1s.")
-                
+            try:
+                if not pd.api.types.is_integer_dtype(parcel_df[column]):
+                    issues_found.append(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
+                if not parcel_df[column].isin([0, 1]).all():
+                    issues_found.append(f"Column '{column}' should only contain 0s or 1s.")
+            except Exception as e:
+                issues_found.append(f"Error checking 'covered' column: {str(e)}")
         else:
-            type_matched = False
-            for expected_type_str, type_name in expected_types:
-                if expected_type_str == "string":
-                    if pd.api.types.is_string_dtype(parcel_df[column]) or actual_type == "object":
-                        type_matched = True
-                        break
-                else:
-                    if pd.api.types.is_dtype_equal(actual_type, expected_type_str):
-                        type_matched = True
-                        break
-            
-            if not type_matched:
-                issues_found.append(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
-    
-    if issues_found != []:       
-        raise DataFrameValidationError("\n" + "\n".join(issues_found))        
+            # Check data type for other columns
+            try:
+                type_matched = False
+                for expected_type_str, type_name in expected_types:
+                    if expected_type_str == "string":
+                        if pd.api.types.is_string_dtype(parcel_df[column]) or actual_type == "object":
+                            type_matched = True
+                            break
+                    else:
+                        if pd.api.types.is_dtype_equal(actual_type, expected_type_str):
+                            type_matched = True
+                            break
+                
+                if not type_matched:
+                    issues_found.append(f"Column '{column}' has incorrect data type. Expected {expected_types}, got {actual_type}.")
+            except Exception as e:
+                issues_found.append(f"Error checking data type of column '{column}': {str(e)}")
+
+    # Always reach this point and raise the custom exception if issues are found
+    if issues_found:
+        raise DataFrameValidationError("\n" + "\n".join(issues_found))
 
     return True
